+++ src/meta/CMakeLists.txt
@@ -93,9 +93,9 @@ nebula_add_library(
 
 nebula_add_library(
     meta_http_handler OBJECT
-    MetaHttpIngestHandler.cpp
-    MetaHttpDownloadHandler.cpp
-    MetaHttpReplaceHostHandler.cpp
+    http/MetaHttpIngestHandler.cpp
+    http/MetaHttpDownloadHandler.cpp
+    http/MetaHttpReplaceHostHandler.cpp
 )
 
 nebula_add_library(
