+++ src/meta/http/MetaHttpDownloadHandler.cpp
