+++ src/meta/http/MetaHttpDownloadHandler.h
@@ -4,11 +4,12 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#ifndef META_METAHTTPINGESTHANDLER_H
-#define META_METAHTTPINGESTHANDLER_H
+#ifndef META_HTTP_METAHTTPDOWNLOADHANDLER_H_
+#define META_HTTP_METAHTTPDOWNLOADHANDLER_H_
 
 #include "common/base/Base.h"
 #include "webservice/Common.h"
+#include "common/hdfs/HdfsHelper.h"
 #include "common/thread/GenericThreadPool.h"
 #include "kvstore/KVStore.h"
 #include <proxygen/httpserver/RequestHandler.h>
@@ -18,11 +19,12 @@ namespace meta {
 
 using nebula::HttpCode;
 
-class MetaHttpIngestHandler : public proxygen::RequestHandler {
+class MetaHttpDownloadHandler : public proxygen::RequestHandler {
 public:
-    MetaHttpIngestHandler() = default;
+    MetaHttpDownloadHandler() = default;
 
     void init(nebula::kvstore::KVStore *kvstore,
+              nebula::hdfs::HdfsHelper *helper,
               nebula::thread::GenericThreadPool *pool);
 
     void onRequest(std::unique_ptr<proxygen::HTTPMessage> headers) noexcept override;
@@ -37,17 +39,23 @@ public:
 
     void onError(proxygen::ProxygenError error) noexcept override;
 
-    bool ingestSSTFiles(GraphSpaceID space);
+private:
+    bool dispatchSSTFiles(const std::string& host,
+                          int32_t port,
+                          const std::string& path);
 
 private:
     HttpCode err_{HttpCode::SUCCEEDED};
-    GraphSpaceID space_;
+    std::string hdfsHost_;
+    int32_t hdfsPort_;
+    std::string hdfsPath_;
+    GraphSpaceID spaceID_;
     nebula::kvstore::KVStore *kvstore_;
+    nebula::hdfs::HdfsHelper *helper_;
     nebula::thread::GenericThreadPool *pool_;
 };
 
 }  // namespace meta
 }  // namespace nebula
 
-#endif  // META_METAHTTPINGESTHANDLER_H
-
+#endif  // META_HTTP_METAHTTPDOWNLOADHANDLER_H_
